Add support for IPv6 (ip6tables)
Introduces a Configurator abstraction for decoupling rule generation
from the binaries used and provides IPv4 and IPv6 configurator
constructors.

Whether to use the IPv4 or IPv6 configurator is determined dynamically
by inspecting local IPs for presence of an IPv4 and/or IPv6 address.
Go 1.18's `net/netip` is used, so all edge cases should be covered when
parsing IPv6.